+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.7.1",
   "description": "Node.js native bindings to FFmpeg.",
   "main": "ts/index.js",
-  "types": "types.d.ts",
+  "types": "ts/index.d.ts",
   "scripts": {
     "preinstall": "node ts/install_ffmpeg.js",
     "install": "node-gyp rebuild",
