+++ ts/types/Beamstreams.d.ts
