+++ ts/types/Frame.d.ts
