+++ ts/types/Governor.d.ts
