+++ ts/types/PrivClass.d.ts
