+++ ts/types/Stream.d.ts
